# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Logenhelfer API
  version: 1.0.0
  description: Backend-agnostic contract. Rails serves it today, Express later.
paths:
...
  /api/v1/session:
    post:
      operationId: createSession
      description: >-
        In addition to the per-IP rate limit (see the shared 429 response),
        repeated consecutive failed attempts against the same email lock
        that email out for a period that grows with each further failure,
        independent of source IP. A locked-out attempt returns the identical
        401 `invalid_credentials` body as an ordinary wrong-password
        response - this endpoint never reveals lockout state to the caller.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: >-
            Password verified. The exact body depends on the account's MFA
            state (see mfaLoginResult): a full login (refresh cookie set,
            `access_token`+`user` present) when no MFA method is enrolled
            (or the enrolled-but-unconfigured mandatory-mode grace period
            hasn't ended, or the request carries a trusted-device cookie
            already covering this account); `{ setup_required: true }`
            with no cookie/tokens when MFA is mandatory, the grace period
            has ended, and the user has no method enrolled yet; or
            `{ mfa_pending_token }` with no cookie when a verified method
            exists and must be challenged via
            `/api/v1/mfa/challenge/verify` before a real session is
            issued. The server sets a `refresh_token` cookie scoped to
            path `/api/v1/session`, marked `HttpOnly`, `Secure` (in
            production), `SameSite=Strict`, with a 30-day expiry, only for
            the full-login variant. Any reimplementation of this endpoint
            (e.g. an Express port) must reproduce this cookie contract
            exactly, or the frontend's refresh flow will break silently.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MfaLoginResult' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    delete:
      operationId: deleteSession
      security: []
      responses:
        '204': { description: Logged out; refresh family revoked }
  /api/v1/session/refresh:
    post:
      operationId: refreshSession
      security: []
      responses:
        '200':
          description: >-
            Rotated; new access token. Reads the `refresh_token` cookie
            (name `refresh_token`, path `/api/v1/session`, `HttpOnly`,
            `Secure` in production, `SameSite=Strict`, 30-day TTL) and
            re-sets it with the same attributes on rotation. This cookie
            contract is the backend-coupled seam of the auth flow and must
            be preserved byte-for-byte by any alternative backend
            implementation.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionPayload' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /api/v1/me:
    get:
      operationId: getMe
      responses:
        '200':
          description: Current user and abilities
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [user, abilities, mfa_setup_required]
                properties:
                  user: { $ref: '#/components/schemas/MeUser' }
                  abilities:
                    type: object
                    additionalProperties:
                      type: array
                      items: { type: string, enum: [read, create, update, destroy, index, user_stats, downloads, file_stats, user_file_stats, mem_stats, names_list] }
                  mfa_setup_required: { type: boolean }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/events:
    get:
      operationId: listEvents
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
        - name: sort
          in: query
          schema: { type: string, default: date }
        - name: from
          in: query
          description: Inclusive lower bound on event date (YYYY-MM-DD).
          schema: { type: string, format: date }
        - name: to
          in: query
          description: Inclusive upper bound on event date (YYYY-MM-DD).
          schema: { type: string, format: date }
      responses:
        '200':
          description: Paginated event list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventList' }
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createEvent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EventInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Event' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /api/v1/external_events:
    get:
      operationId: listExternalEvents
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
        - name: from
          in: query
          description: Inclusive lower bound on event date (YYYY-MM-DD). When omitted, defaults to "yesterday onward".
          schema: { type: string, format: date }
        - name: to
          in: query
          description: Inclusive upper bound on event date (YYYY-MM-DD).
          schema: { type: string, format: date }
      responses:
        '200':
          description: Paginated external event list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExternalEventList' }
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createExternalEvent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExternalEventInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExternalEvent' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /api/v1/external_event_ics_sources:
    get:
      operationId: listExternalEventIcsSources
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
        - name: sort
          in: query
          schema: { type: string, default: name }
      responses:
        '200':
          description: Paginated ICS source list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExternalEventIcsSourceList' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createExternalEventIcsSource
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExternalEventIcsSourceInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExternalEventIcsSource' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /api/v1/members:
    get:
      operationId: listMembers
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
        - name: sort
          in: query
          schema: { type: string, default: lastname }
        - name: search
          in: query
          allowEmptyValue: true
          schema: { type: string }
      responses:
        '200':
          description: Paginated, admin-visibility-filtered member list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MemberList' }
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createMember
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MemberInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Member' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /api/v1/seekers:
    get:
      operationId: listSeekers
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
        - name: sort
          in: query
          schema: { type: string, default: lastname }
        - name: filter
          in: query
          schema: { type: string, enum: [active, accepted, inactive, declined], default: active }
      responses:
        '200':
          description: Paginated seeker list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SeekerList' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createSeeker
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SeekerInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Seeker' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /api/v1/roles:
    get:
      operationId: listRoles
      parameters:
        - name: scope
          in: query
          schema: { type: string, enum: [positions, administrational] }
      responses:
        '200':
          description: All roles (small, unpaginated reference list), or only officer-eligible positions when scope=positions
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoleList' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/categories:
    get:
      operationId: listCategories
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
        - name: sort
          in: query
          schema: { type: string, default: name }
      responses:
        '200':
          description: Paginated category list, filtered to categories visible to the caller
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoryList' }
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoryInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /api/v1/directories:
    get:
      operationId: listDirectories
      parameters:
        - name: category_slug
          in: query
          required: true
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
      responses:
        '200':
          description: Paginated directory list for one category, filtered to directories visible to the caller
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DirectoryList' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: createDirectory
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DirectoryInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Directory' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: category_slug did not match any category
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /api/v1/attached_files:
    get:
      operationId: listAttachedFiles
      parameters:
        - name: directory_slug
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Files in a directory the caller is permitted to see
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AttachedFileList' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createAttachedFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, directory_slug]
              properties:
                file: { type: string, format: binary }
                directory_slug: { type: string }
                # Deliberately untyped (no `type: array`), unlike the PATCH
                # endpoint's JSON-body role_ids below. FileForm/api.ts sends
                # selected roles as repeated `role_ids[]` multipart fields;
                # multer (via append-field) turns even a single occurrence
                # into a one-element array. express-openapi-validator's own
                # multipart handling (openapi.request.validator.js's
                # multipartNested) JSON.parses any multipart field whose
                # declared type is array/object - which for a one-element
                # array coerces to its string join first (['5'] -> '5'), then
                # parses that as the bare number 5, silently replacing the
                # array with a scalar and failing "role_ids must be array"
                # for exactly one selected role. Leaving this untyped skips
                # that coercion entirely; attachedFiles.ts's
                # extractMultipartRequest already validates/coerces whatever
                # shape multer hands it (array of numeric strings) by hand.
                role_ids: {}
      responses:
        '201':
          description: Uploaded file
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AttachedFile' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /api/v1/districts:
    get:
      operationId: listDistricts
      responses:
        '200':
          description: All districts (small, unpaginated reference list)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DistrictList' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createDistrict
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DistrictInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DistrictSummary' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /api/v1/academic_titles:
    get:
      operationId: listAcademicTitles
      responses:
        '200':
          description: All academic titles (small, unpaginated reference list)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AcademicTitleList' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createAcademicTitle
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AcademicTitleInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AcademicTitleSummary' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /api/v1/lodges:
    get:
      operationId: listLodges
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
        - name: sort
          in: query
          schema: { type: string, default: name }
      responses:
        '200':
          description: Paginated lodge list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LodgeList' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createLodge
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LodgeInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lodge' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /api/v1/officers:
    get:
      operationId: listOfficers
      parameters:
        - name: lodge_slug
          in: query
          required: true
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
      responses:
        '200':
          description: Paginated officer list for one lodge
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OfficerList' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: createOfficer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OfficerInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Officer' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: lodge_slug did not match any lodge
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /api/v1/announcements:
    get:
      operationId: listAnnouncements
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
        - name: sort
          in: query
          schema: { type: string, default: '-created_at' }
      responses:
        '200':
          description: Paginated announcement list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnnouncementList' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createAnnouncement
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AnnouncementInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Announcement' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /api/v1/app_config:
    get:
      operationId: getAppConfig
      responses:
        '200':
          description: Every known AppConfig setting and its current value
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppConfigValues' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    patch:
      operationId: updateAppConfig
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AppConfigValues' }
      responses:
        '200':
          description: The full, updated set of AppConfig settings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppConfigValues' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error: { type: string }
    BadRequest:
      description: Malformed or missing request parameters
      content:
        application/json:
          schema:
            type: object
            required: [error, detail]
            properties:
              error: { type: string }
              detail: { type: string }
...
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error: { type: string }
...
  schemas:
...
    MeUser:
      type: object
      additionalProperties: false
      required: [id, uuid, email, firstname, lastname, subscribed_to_announcements, gdpr_accepted]
      properties:
        id: { type: integer }
        uuid: { type: string }
        email: { type: string }
        firstname: { type: string, nullable: true }
        lastname: { type: string, nullable: true }
        subscribed_to_announcements: { type: boolean }
        gdpr_accepted: { type: boolean }
    SessionPayload:
      type: object
      additionalProperties: false
      required: [access_token, user]
      properties:
        access_token: { type: string }
        user: { $ref: '#/components/schemas/User' }
...
    MfaLoginResult:
      type: object
      additionalProperties: false
      properties:
        mfa_pending_token: { type: string }
        setup_required: { type: boolean }
        access_token: { type: string }
        user: { $ref: '#/components/schemas/User' }
...
    EventList:
      type: object
      additionalProperties: false
      required: [rows, row_count]
      properties:
        rows:
          type: array
          items: { $ref: '#/components/schemas/Event' }
        row_count: { type: integer }
...
security:
  - bearerAuth: []